refactor(client): extract pagination as a hook to reuse
