+++ client/src/components/Inventory/Parts.jsx
@@ -1,26 +1,17 @@
-import React, { Fragment, useState } from 'react';
-import { Table, Thead, Tbody, Tr, TableCaption, Heading } from '@chakra-ui/react';
-import Loader from 'components/common/Loader';
-import { getParts } from 'utils/api/parts.js';
-import { useQuery } from 'react-query';
-import { StyledTableRow, StyledTableHeader, StyledTableCell } from 'components/common/Table.jsx';
+import { Heading, Table, TableCaption, Tbody, Thead, Tr } from '@chakra-ui/react';
 import { TablePagination } from '@material-ui/core';
-import { NoResultImage } from 'components/common/Image.jsx';
 import ExportFiles from 'components/common/ExportFiles.jsx';
+import { NoResultImage } from 'components/common/Image.jsx';
+import Loader from 'components/common/Loader';
+import { StyledTableCell, StyledTableHeader, StyledTableRow } from 'components/common/Table.jsx';
+import usePagination from 'hooks/usePagination.jsx';
+import { Fragment } from 'react';
+import { useQuery } from 'react-query';
+import { getParts } from 'utils/api/parts.js';
 
 const Parts = () => {
   const { isLoading, isSuccess, data } = useQuery('parts', getParts);
-  const [page, setPage] = useState(0);
-  const [rowsPerPage, setRowsPerPage] = useState(10);
-
-  const handleChangePage = (event, newPage) => {
-    setPage(newPage);
-  };
-
-  const handleChangeRowsPerPage = (event) => {
-    setRowsPerPage(parseInt(event.target.value, 10));
-    setPage(0);
-  };
+  const { handleChangePage, handleChangeRowsPerPage, page, rowsPerPage } = usePagination();
 
   if (isLoading) {
     return <Loader />;
