+++ client/src/hooks/usePagination.jsx
@@ -1,0 +1,19 @@
+import { useState } from 'react';
+
+const usePagination = () => {
+  const [page, setPage] = useState(0);
+  const [rowsPerPage, setRowsPerPage] = useState(10);
+
+  const handleChangePage = (event, newPage) => {
+    setPage(newPage);
+  };
+
+  const handleChangeRowsPerPage = (event) => {
+    setRowsPerPage(parseInt(event.target.value, 10));
+    setPage(0);
+  };
+
+  return { handleChangePage, handleChangeRowsPerPage, page, rowsPerPage };
+};
+
+export default usePagination;
