+++ client/src/hooks/useQualityPartsTable.jsx
@@ -5,19 +5,8 @@ import { deleteDefects, getDefects } from 'utils/api/defect.js';
 
 const useQualityPartsTable = () => {
   const { isLoading, isSuccess, data, refetch } = useQuery('defects', getDefects);
-  const [page, setPage] = useState(0);
-  const [rowsPerPage, setRowsPerPage] = useState(10);
   const [selected, setSelected] = useState([]);
   const toast = useToast();
-
-  const handleChangePage = (e, newPage) => {
-    setPage(newPage);
-  };
-
-  const handleChangeRowsPerPage = (e) => {
-    setRowsPerPage(parseInt(e.target.value, 10));
-    setPage(0);
-  };
 
   const handleStatusColor = (status) => {
     switch (status) {
@@ -86,8 +75,6 @@ const useQualityPartsTable = () => {
   };
 
   return {
-    handleChangePage,
-    handleChangeRowsPerPage,
     handleStatusColor,
     isSelected,
     handleSelectAllClick,
@@ -95,8 +82,6 @@ const useQualityPartsTable = () => {
     handleDelete,
     isLoading,
     isSuccess,
-    page,
-    rowsPerPage,
     data,
     selected,
   };
